WIP: Moved Device to utils package
