+++ .gitignore
@@ -11,7 +11,7 @@ dist/
 typings
 !/typings
 .cache
-!packages/fiori3/src/utils/Device/lib
+!packages/utils/src/Device/lib
 !packages/fiori3/src/lib
 build
 /packages/*/npm/**/*.d.ts
