--- packages/fiori3/src/lib/Device.ts
@@ -1,3 +1,0 @@
-import { Device } from '../utils/Device';
-
-export { Device };
