+++ packages/utils/src/Device/Browser.test.ts
