+++ packages/utils/src/Device/EventRegistry.test.ts
