+++ packages/utils/src/Device/Media.test.ts
@@ -1,7 +1,7 @@
 import { Media, RANGESETS } from './Media';
 import { EventRegistry } from './EventRegistry';
 import * as sinon from 'sinon';
-import { setupMatchMedia } from '../../../../../config/jestsetup';
+import { setupMatchMedia } from '../../../../config/jestsetup';
 
 const defaultSupportInstance = {
   matchmedialistener: false,
