+++ packages/utils/src/Device/lib/modernizr-config.json
