+++ packages/utils/src/Device/lib/modernizr.js
