+++ packages/utils/src/Device/utils.ts
