+++ packages/utils/src/isNumeric.ts
